requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.34.0,<1.0
gunicorn>=23.0.0,<24.0
litellm>=1.50.0,<2.0
yfinance>=0.2.0,<1.0
finnhub-python>=2.4.0,<3.0
pandas>=2.2.0,<3.0
numpy>=1.26.0,<3.0
matplotlib>=3.9.0,<4.0
mplfinance>=0.12.0,<1.0
apscheduler>=3.10.0,<4.0
pydantic-settings>=2.5.0,<3.0
httpx>=0.27.0,<1.0
sqlalchemy>=2.0.0,<3.0
pytz
